Configure and style DCF tabs
